Moved enumerations into separate folder
