+++ include/compute/coupled_interfaces.hpp
@@ -1,10 +1,10 @@
 #ifndef _COMPUTE_COUPLED_INTERFACES_HPP
 #define _COMPUTE_COUPLED_INTERFACES_HPP
 
+#include "enumerations/interface.hpp"
 #include "kokkos_abstractions.h"
 #include "macros.hpp"
 #include "mesh/coupled_interfaces/coupled_interfaces.hpp"
-#include "specfem_enums.hpp"
 
 namespace specfem {
 namespace compute {
