+++ include/domain/impl/elements/elastic/elastic2d_isotropic.tpp
@@ -6,7 +6,7 @@
 #include "domain/impl/elements/element.hpp"
 #include "globals.h"
 #include "kokkos_abstractions.h"
-#include "specfem_enums.hpp"
+#include "enumerations/interface.hpp"
 #include "specfem_setup.hpp"
 #include <Kokkos_Core.hpp>
