+++ include/domain/impl/receivers/acoustic/acoustic2d.hpp
@@ -3,7 +3,7 @@
 
 #include "constants.hpp"
 #include "domain/impl/receivers/receiver.hpp"
-#include "specfem_enums.hpp"
+#include "enumerations/interface.hpp"
 #include <Kokkos_Core.hpp>
 
 namespace specfem {
