+++ include/domain/impl/receivers/kernel.hpp
@@ -3,9 +3,9 @@
 
 #include "domain/impl/receivers/acoustic/interface.hpp"
 #include "domain/impl/receivers/elastic/interface.hpp"
+#include "enumerations/interface.hpp"
 #include "kokkos_abstractions.h"
 #include "quadrature/interface.hpp"
-#include "specfem_enums.hpp"
 #include "specfem_setup.hpp"
 
 namespace specfem {
