+++ include/domain/impl/sources/acoustic/acoustic2d_isotropic.tpp
@@ -7,7 +7,7 @@
 #include "globals.h"
 #include "kokkos_abstractions.h"
 #include "source_time_function/source_time_function.hpp"
-#include "specfem_enums.hpp"
+#include "enumerations/interface.hpp"
 #include "specfem_setup.hpp"
 #include <Kokkos_Core.hpp>
