+++ include/domain/impl/sources/kernel.tpp
@@ -6,7 +6,7 @@
 #include "domain/impl/sources/elastic/interface.hpp"
 #include "kernel.hpp"
 #include "kokkos_abstractions.h"
-#include "specfem_enums.hpp"
+#include "enumerations/interface.hpp"
 #include "specfem_setup.hpp"
 #include <Kokkos_Core.hpp>
