+++ include/enumerations/interface.hpp
@@ -1,0 +1,10 @@
+#ifndef _ENUMERATIONS_INTERFACE_HPP_
+#define _ENUMERATIONS_INTERFACE_HPP_
+
+#include "dimension.hpp"
+#include "medium.hpp"
+#include "properties.hpp"
+#include "quadrature.hpp"
+#include "specfem_enums.hpp"
+
+#endif /* _ENUMERATIONS_INTERFACE_HPP_ */
