+++ include/material/material.hpp
@@ -2,7 +2,7 @@
 #define _MATERIAL_HPP
 
 #include "constants.hpp"
-#include "specfem_enums.hpp"
+#include "enumerations/interface.hpp"
 #include "specfem_mpi/interface.hpp"
 #include "specfem_setup.hpp"
 #include "utilities/interface.hpp"
