+++ include/receiver/receiver.hpp
@@ -2,9 +2,9 @@
 #define _RECEIVER_HPP
 
 #include "constants.hpp"
+#include "enumerations/interface.hpp"
 #include "kokkos_abstractions.h"
 #include "quadrature/interface.hpp"
-#include "specfem_enums.hpp"
 #include "specfem_mpi/interface.hpp"
 #include "specfem_setup.hpp"
 #include <cmath>
