+++ include/source/force_source.hpp
@@ -1,11 +1,11 @@
 #ifndef _FORCE_SOURCE_HPP
 #define _FORCE_SOURCE_HPP
 
+#include "enumerations/interface.hpp"
 #include "kokkos_abstractions.h"
 #include "quadrature/interface.hpp"
 #include "source.hpp"
 #include "source_time_function/interface.hpp"
-#include "specfem_enums.hpp"
 #include "specfem_mpi/interface.hpp"
 #include "specfem_setup.hpp"
 #include "utilities/interface.hpp"
